cleaned some code and removed unecessary saga
I originally created a saga thinking that we needed to grab activity data. I quickly realized we already had a prefiltered list of activities per organization. I removed it now.

Aside from that, file submission was spruced up a bit
added a new route for operators (and admin) to expose the new table.